+++ src/helpers/index.ts
@@ -3,6 +3,7 @@
  * Reusable utility classes for enterprise applications
  */
 
+export * from './messages';
 export * from './CurrencyHelper';
 export * from './DateHelper';
 export * from './ProgressHelper';